--- .vscode/settings.json
@@ -1,2 +1,0 @@
-{
-}
