+++ docs/기능 목록.md
@@ -22,10 +22,6 @@
 - [x] sass => style component
 - [] skeleton 도입
 
-#### 배포
-
-- [] netlify 배포 브랜치 변경
-
 ---
 
 ### 개발 진행 순서
@@ -68,8 +64,6 @@
 ---
 
 ### 회고
-
-- 테스트 방법에 env 추가
 
 ---
