+++ src/models/order.ts
@@ -1,2 +1,1 @@
-// TODO: 영문 삭제
-export type Order = "최신순" | "좋아요순" | "recent" | "favorite";
+export type Order = "최신순" | "좋아요순";
