+++ src/pages/market/FleaMarket.tsx
@@ -1,19 +1,19 @@
 import GNB from "components/Navbar";
 import MainLayout from "components/Layout";
 import Pagination from "components/Pagination";
-import { AllProducts, BestProducts } from "templates/MarketMain";
-import MarketMainProvider from "contexts/MarketMain";
+import { AllProducts, BestProducts } from "templates/FleaMarket";
+import FleaMarketProvider from "contexts/react-context/FleaMarket";
 
-export default function MarketMainPage() {
+export default function FleaMarket() {
   return (
     <>
       <GNB />
       <MainLayout>
-        <MarketMainProvider>
+        <FleaMarketProvider>
           <BestProducts />
           <AllProducts />
           <Pagination />
-        </MarketMainProvider>
+        </FleaMarketProvider>
       </MainLayout>
     </>
   );
